Put test in file w/ descriptive name
...and fix test assertion
